# Small optimization around makeError