--- a/core/shared/src/main/scala/cats/parse/Parser.scala
+++ b/core/shared/src/main/scala/cats/parse/Parser.scala
@@ -1802,8 +1802,16 @@ object Parser extends ParserInstances {
 
       override def toString = s"CharIn($min, bitSet = ..., $ranges)"
 
-      def makeError(offset: Int): Chain[Expectation] =
-        Chain.fromSeq(ranges.toList.map { case (s, e) => Expectation.InRange(offset, s, e) })
+      def makeError(offset: Int): Chain[Expectation] = {
+        var result = Chain.empty[Expectation]
+        var aux = ranges.toList
+        while (aux.nonEmpty) {
+          val (s, e) = aux.head
+          result = result :+ Expectation.InRange(offset, s, e)
+          aux = aux.tail
+        }
+        result
+      }
 
       override def parseMut(state: State): Char = {
         val offset = state.offset